Setup for A* search

+++ src/Structure.cpp
@@ -6,7 +6,7 @@ Structure::Structure(Rectangle body,std::map<Vector2,Chunk*,Vec2Compare>* map){
   
   for(int i = modChunkLength(body.x); i < modChunkLength(body.x+body.width); i = i+chunkLength){
     for(int j = modChunkLength(body.y); j < modChunkLength(body.y + body.height); j = j+chunkLength){
-      map->find(Vector2{i,j})->second->setChunkType(structSpace);
+      map->find(Vector2{(float)i,(float)j})->second->setChunkType(structSpace);
     }
   }
 }
